ScoopCloudFunctions/functions/requirements.txt
firebase-functions
firebase-admin
google-cloud-tasks>=2.13.0